requirements.txt
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.10
django-allauth==0.32.0
django-braces==1.11.0
django-countries==4.0
django-geoposition==0.3.0
django-oauth-toolkit==1.0.0
django-resized==0.3.5
django-rest-auth==0.9.1
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.6.3
geoip2==2.4.2
idna==2.5
Jinja2==2.8
MarkupSafe==1.0
maxminddb==1.3.0
mysqlclient==1.3.10
oauthlib==2.0.2
olefile==0.44
Pillow==3.4.2
PyJWT==1.5.2
python3-openid==3.1.0
pytz==2017.2
requests==2.17.3
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.21.1
